kiml-api/src/main.py
from fastapi import FastAPI
from fastapi.staticfiles import StaticFiles
from starlette.middleware.cors import CORSMiddleware
from dotenv import load_dotenv

from routers import experiment, login, status, dashboard


load_dotenv()
app = FastAPI()

origins = [
    "http:///127.0.0.1:12345"
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(experiment.router)
app.include_router(login.router)
app.include_router(status.router)
app.include_router(dashboard.router)

app.mount("/static", StaticFiles(directory="static"), name="static")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app=app, host="0.0.0.0", port=12345)