kiml-api/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.1
httpx==0.24.1
python-dotenv==1.0.0
--extra-index-url https://pypi.ml.kakaoicloud-kr-gov.com/simple --trusted-host pypi.ml.kakaoicloud-kr-gov.com 
kiml==1.2.22
jinja2==3.1.2
sqlalchemy[asyncio]==2.0.19
aioredis==2.0.1